pyproject.toml
[project]
name = "strands-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.0",
    "strands-agents[otel]>=0.1.9",
    "strands-agents-builder>=0.1.4",
    "streamlit>=1.46.1",
]
